Fix Zotero.RDF.getResourceURI() for nodes with rdf:about and rdf:value
Return a URI based on the `rdf:about` attribute if present rather than
the contents of the `rdf:value` element. In particular, this means that
for an exported note (a `bib:Memo` node), a URI ending in, e.g.,
`#item_3` will be returned from `getResourceURI()` rather than the note
cnotents.

This is part of the fix for standalone notes in RDF exports not being
added to collections they were in on re-import, due to the lack of a
valid id.

This behavior dates to zotero/zotero#453 from 2014 (which I apparently
knew broke `getResourceURI()` for notes as of zotero/zotero@9c2d0d727 in
2019).