Translator tester: Be more vigilant against empty test cases.
The test cases in a translator script may be empty, not just with an
empty array for `var testCases`, but in the sense of completely missing.

In that case, don't pass invalid JSON (all whitespaces or empty string)
to `JSON.parse()`, and use an empty array instead.

This fixes an annoying `SyntaxError` in the console while loading the
file `Ovid.js` in the translators repository. That translator should be
fixed, too.